requirements.txt
django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2
django-extensions==3.2.3
python-dotenv==1.0.1
python-decouple==3.8
psycopg2-binary==2.9.9
requests==2.32.3
debugpy==1.8.5
